+++ CHANGELOG.md
@@ -6,6 +6,10 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 0.2.0 - 9/13/17 Weds
+
+* [FEATURE] Module to convert nested (and non-nested) map keys to "stringified" keys
+
 ## 0.1.0 - 9/12/17 Tues
 
-* [FEATURE] Module to convert nested (and non-nested) map keys to atoms
+* [FEATURE] Module to convert nested (and non-nested) map keys to "atomized" keys